Fix lai sau khi thiet ke lai CSDL 1

+++ app/Http/Controllers/Admin/RolesController.php
@@ -5,14 +5,16 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Role;
+use App\Models\PermissionRole;
+use App\Http\Resources\RoleResource;
 
 class RolesController extends Controller
 {
 
     public function index()
     {
-        $roles = Role::orderBy('created_at', 'DESC')->get();
-        return response()->json($roles); 
+        $roles = Role::with('permission_role.permission')->orderBy('created_at', 'DESC')->get();
+        return response()->json(RoleResource::collection($roles));
     }
 
     public function store(Request $request)
@@ -21,39 +23,60 @@ class RolesController extends Controller
         $role->name = $request['name'];
         $role->description = $request['description'];
         $role->save();
+
+        $getPermissions = $request['permission_id'];      
+
+        // create table permission_role
+        foreach($getPermissions as $item) {
+            $permissionRole = new PermissionRole;
+            $permissionRole->role_id = $role->id;
+            $permissionRole->permission_id = $item;
+            $permissionRole->save();
+        }
+
         return response()->json('Role created!');
     }
 
     public function show($id)
     {
-        $role = Role::find($id);
-        return response()->json($role);
+        $role = Role::with('permission_role.permission')->find($id);
+        return response()->json(new RoleResource($role));
     }
 
     public function update(Request $request, $id)
     {
         $role = Role::find($id);
         $role->update($request->all());
+
+        $getPermissions = $request['permission_id'];
+        
+        // delte permission of role if existed
+        $permisisonsRoles = PermissionRole::where(['role_id' => $id])->get();
+        foreach($permisisonsRoles as $role) {
+            PermissionRole::where(['role_id' => $id])->delete();
+        }
+
+        // create table permission_role
+        foreach($getPermissions as $item) {
+            $permissionRole = new PermissionRole;
+            $permissionRole->role_id = $id;
+            $permissionRole->permission_id = $item;
+            $permissionRole->save();
+        }
+    
         return response()->json('Role updated!');
     }
 
-    public function destroy($id)
+    public function destroyIds(Request $request)
     {
-        $role = Role::find($id);
-        $role->delete();
-        return response()->json(['success'=>'true'], 200);
-    }
-
-    public function destroyAll()
-    {
-        $roles = Role::all();
-        
-        foreach($roles as $key => $value) {
-            $role = Role::find($roles[$key]['id']);
+        $selectedIds = $request->all(); 
+        $roles = Role::whereIn('id', $selectedIds)->get();  
+        foreach($roles as $role) {
+            PermissionRole::where(['role_id' => $role->id])->delete();
             $role->delete();
-        }
+        }      
         return response()->json([
-            'status' => false,
+            'success' => true,
             'message' => "Deleted All."
         ], 200);
     }
